Moved long route to dedicated knowledge_panel file + utils to avoid circular dependency
tests passed with pytest command
